Made that slightly more generic, so it applies to other platforms which implement restoreWifiAsync (Kobo, Cervantes, PRSTux)
